@@ -6,27 +6,22 @@ description: Demonstrates how to use the mdsflutter plugin.
 publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
 
-  mdsflutter:
-    # When depending on this package from a real application you should use:
-    #   mdsflutter: ^x.y.z
-    # See https://dart.dev/tools/pub/dependencies#version-constraints
-    # The example app is bundled with the plugin so we use a path dependency on
-    # the parent directory to use the current plugin's version.
-    path: ../
+  mdsflutter: 
+    path: ../../mdsflutter
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
-  cupertino_icons: ^0.1.3
+  cupertino_icons: ^1.0.3
 
-  provider: ^3.0.0
+  provider: ^6.0.1
 
-  permission_handler: ^5.0.1
+  permission_handler: ^8.2.2
 
 dev_dependencies:
   flutter_test:
