@@ -1,17 +1,16 @@
 name: mdsflutter
 description: Flutter plugin for MDS (Movesense Device Service) that is used for communicating with Movesense devices.
-version: 1.0.0
-author: Tugberk Akdogan
+version: 1.1.0
 homepage: https://github.com/tugberka/mdsflutter
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
-  flutter: ">=1.10.0"
+  sdk: '>=2.12.0 <3.0.0'
+  flutter: ">=2.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
-  protobuf: any
+  protobuf: ^2.0.0
 
 dev_dependencies:
   flutter_test:
