@@ -18,13 +18,13 @@ class DeviceInteractionWidget extends StatefulWidget {
 }
 
 class _DeviceInteractionWidgetState extends State<DeviceInteractionWidget> {
-  AppModel _appModel;
+  AppModel? _appModel;
 
   @override
   void initState() {
     super.initState();
     _appModel = Provider.of<AppModel>(context, listen: false);
-    _appModel.onDeviceMdsDisconnected((device) => { Navigator.pop(context) });
+    _appModel?.onDeviceMdsDisconnected((device) => {Navigator.pop(context)});
   }
 
   void _onAccelerometerButtonPressed(DeviceModel deviceModel) {
@@ -45,7 +45,7 @@ class _DeviceInteractionWidgetState extends State<DeviceInteractionWidget> {
 
   @override
   void dispose() {
-    _appModel.disconnectFromDevice(widget.device);
+    _appModel?.disconnectFromDevice(widget.device);
     super.dispose();
   }
 
@@ -53,12 +53,12 @@ class _DeviceInteractionWidgetState extends State<DeviceInteractionWidget> {
   Widget build(BuildContext context) {
     Device device = widget.device;
     return ChangeNotifierProvider(
-      create: (context) => DeviceModel(device.name, device.serial),
+      create: (context) => DeviceModel(device.name ?? "", device.serial ?? ""),
       child: Consumer<DeviceModel>(
         builder: (context, model, child) {
           return Scaffold(
               appBar: AppBar(
-                title: Text(device.name),
+                title: Text(device.name ?? ""),
               ),
               body: Column(
                 mainAxisSize: MainAxisSize.min,
@@ -68,8 +68,7 @@ class _DeviceInteractionWidgetState extends State<DeviceInteractionWidget> {
                   _ledItem(model),
                   _temperatureItem(model)
                 ],
-              )
-          );
+              ));
         },
       ),
     );
@@ -80,8 +79,10 @@ class _DeviceInteractionWidgetState extends State<DeviceInteractionWidget> {
       child: ListTile(
         title: Text("Accelerometer"),
         subtitle: Text(deviceModel.accelerometerData),
-        trailing: RaisedButton(
-          child: Text(deviceModel.accelerometerSubscribed ? "Unsubscribe" : "Subscribe"),
+        trailing: ElevatedButton(
+          child: Text(deviceModel.accelerometerSubscribed
+              ? "Unsubscribe"
+              : "Subscribe"),
           onPressed: () => _onAccelerometerButtonPressed(deviceModel),
         ),
       ),
@@ -93,7 +94,7 @@ class _DeviceInteractionWidgetState extends State<DeviceInteractionWidget> {
       child: ListTile(
         title: Text("Heart rate"),
         subtitle: Text(deviceModel.hrData),
-        trailing: RaisedButton(
+        trailing: ElevatedButton(
           child: Text(deviceModel.hrSubscribed ? "Unsubscribe" : "Subscribe"),
           onPressed: () => _onHrButtonPressed(deviceModel),
         ),
@@ -107,7 +108,7 @@ class _DeviceInteractionWidgetState extends State<DeviceInteractionWidget> {
         title: Text("Led"),
         trailing: Switch(
           value: deviceModel.ledStatus,
-          onChanged: (b) => { deviceModel.switchLed() },
+          onChanged: (b) => {deviceModel.switchLed()},
         ),
       ),
     );
@@ -118,7 +119,7 @@ class _DeviceInteractionWidgetState extends State<DeviceInteractionWidget> {
       child: ListTile(
         title: Text("Temperature"),
         subtitle: Text(deviceModel.temperature),
-        trailing: RaisedButton(
+        trailing: ElevatedButton(
           child: Text("Get"),
           onPressed: () => deviceModel.getTemperature(),
         ),
