@@ -2,7 +2,7 @@
 //  Generated code. Do not modify.
 //  source: protos/mdsflutter.proto
 //
-// @dart = 2.3
+
 // ignore_for_file: annotate_overrides,camel_case_types,unnecessary_const,non_constant_identifier_names,library_prefixes,unused_import,unused_shown_name,return_of_invalid_type
 
 import 'dart:core' as $core;
@@ -22,7 +22,7 @@ class RequestResult extends $pb.GeneratedMessage {
   factory RequestResult.fromBuffer($core.List<$core.int> i, [$pb.ExtensionRegistry r = $pb.ExtensionRegistry.EMPTY]) => create()..mergeFromBuffer(i, r);
   factory RequestResult.fromJson($core.String i, [$pb.ExtensionRegistry r = $pb.ExtensionRegistry.EMPTY]) => create()..mergeFromJson(i, r);
   RequestResult clone() => RequestResult()..mergeFromMessage(this);
-  RequestResult copyWith(void Function(RequestResult) updates) => super.copyWith((message) => updates(message as RequestResult));
+  RequestResult copyWith(void Function(RequestResult) updates) => super.copyWith((message) => updates(message as RequestResult)) as RequestResult;
   $pb.BuilderInfo get info_ => _i;
   @$core.pragma('dart2js:noInline')
   static RequestResult create() => RequestResult._();
@@ -30,7 +30,7 @@ class RequestResult extends $pb.GeneratedMessage {
   static $pb.PbList<RequestResult> createRepeated() => $pb.PbList<RequestResult>();
   @$core.pragma('dart2js:noInline')
   static RequestResult getDefault() => _defaultInstance ??= $pb.GeneratedMessage.$_defaultFor<RequestResult>(create);
-  static RequestResult _defaultInstance;
+  static RequestResult? _defaultInstance;
 
   @$pb.TagNumber(1)
   $core.int get requestId => $_getIZ(0);
@@ -73,7 +73,7 @@ class RequestError extends $pb.GeneratedMessage {
   factory RequestError.fromBuffer($core.List<$core.int> i, [$pb.ExtensionRegistry r = $pb.ExtensionRegistry.EMPTY]) => create()..mergeFromBuffer(i, r);
   factory RequestError.fromJson($core.String i, [$pb.ExtensionRegistry r = $pb.ExtensionRegistry.EMPTY]) => create()..mergeFromJson(i, r);
   RequestError clone() => RequestError()..mergeFromMessage(this);
-  RequestError copyWith(void Function(RequestError) updates) => super.copyWith((message) => updates(message as RequestError));
+  RequestError copyWith(void Function(RequestError) updates) => super.copyWith((message) => updates(message as RequestError)) as RequestError;
   $pb.BuilderInfo get info_ => _i;
   @$core.pragma('dart2js:noInline')
   static RequestError create() => RequestError._();
@@ -81,7 +81,7 @@ class RequestError extends $pb.GeneratedMessage {
   static $pb.PbList<RequestError> createRepeated() => $pb.PbList<RequestError>();
   @$core.pragma('dart2js:noInline')
   static RequestError getDefault() => _defaultInstance ??= $pb.GeneratedMessage.$_defaultFor<RequestError>(create);
-  static RequestError _defaultInstance;
+  static RequestError? _defaultInstance;
 
   @$pb.TagNumber(1)
   $core.int get requestId => $_getIZ(0);
@@ -123,7 +123,7 @@ class Notification extends $pb.GeneratedMessage {
   factory Notification.fromBuffer($core.List<$core.int> i, [$pb.ExtensionRegistry r = $pb.ExtensionRegistry.EMPTY]) => create()..mergeFromBuffer(i, r);
   factory Notification.fromJson($core.String i, [$pb.ExtensionRegistry r = $pb.ExtensionRegistry.EMPTY]) => create()..mergeFromJson(i, r);
   Notification clone() => Notification()..mergeFromMessage(this);
-  Notification copyWith(void Function(Notification) updates) => super.copyWith((message) => updates(message as Notification));
+  Notification copyWith(void Function(Notification) updates) => super.copyWith((message) => updates(message as Notification)) as Notification;
   $pb.BuilderInfo get info_ => _i;
   @$core.pragma('dart2js:noInline')
   static Notification create() => Notification._();
@@ -131,7 +131,7 @@ class Notification extends $pb.GeneratedMessage {
   static $pb.PbList<Notification> createRepeated() => $pb.PbList<Notification>();
   @$core.pragma('dart2js:noInline')
   static Notification getDefault() => _defaultInstance ??= $pb.GeneratedMessage.$_defaultFor<Notification>(create);
-  static Notification _defaultInstance;
+  static Notification? _defaultInstance;
 
   @$pb.TagNumber(1)
   $core.int get subscriptionId => $_getIZ(0);
@@ -165,7 +165,7 @@ class NotificationError extends $pb.GeneratedMessage {
   factory NotificationError.fromBuffer($core.List<$core.int> i, [$pb.ExtensionRegistry r = $pb.ExtensionRegistry.EMPTY]) => create()..mergeFromBuffer(i, r);
   factory NotificationError.fromJson($core.String i, [$pb.ExtensionRegistry r = $pb.ExtensionRegistry.EMPTY]) => create()..mergeFromJson(i, r);
   NotificationError clone() => NotificationError()..mergeFromMessage(this);
-  NotificationError copyWith(void Function(NotificationError) updates) => super.copyWith((message) => updates(message as NotificationError));
+  NotificationError copyWith(void Function(NotificationError) updates) => super.copyWith((message) => updates(message as NotificationError)) as NotificationError;
   $pb.BuilderInfo get info_ => _i;
   @$core.pragma('dart2js:noInline')
   static NotificationError create() => NotificationError._();
@@ -173,7 +173,7 @@ class NotificationError extends $pb.GeneratedMessage {
   static $pb.PbList<NotificationError> createRepeated() => $pb.PbList<NotificationError>();
   @$core.pragma('dart2js:noInline')
   static NotificationError getDefault() => _defaultInstance ??= $pb.GeneratedMessage.$_defaultFor<NotificationError>(create);
-  static NotificationError _defaultInstance;
+  static NotificationError? _defaultInstance;
 
   @$pb.TagNumber(1)
   $core.int get subscriptionId => $_getIZ(0);
