@@ -1,4 +1,3 @@
-
 import 'dart:collection';
 import 'dart:convert';
 
@@ -11,13 +10,13 @@ import 'package:mdsflutter_example/DeviceConnectionStatus.dart';
 import 'dart:developer' as developer;
 
 class AppModel extends ChangeNotifier {
-
-  final Set<Device> _deviceList = Set();
+  final Set<Device?> _deviceList = Set();
   bool _isScanning = false;
-  void Function(Device) _onDeviceMdsConnectedCb;
-  void Function(Device) _onDeviceDisonnectedCb;
+  void Function(Device)? _onDeviceMdsConnectedCb;
+  void Function(Device)? _onDeviceDisonnectedCb;
 
-  UnmodifiableListView<Device> get deviceList => UnmodifiableListView(_deviceList);
+  UnmodifiableListView<Device?> get deviceList =>
+      UnmodifiableListView(_deviceList);
 
   bool get isScanning => _isScanning;
 
@@ -33,7 +32,8 @@ class AppModel extends ChangeNotifier {
 
   void startScan() {
     _deviceList.forEach((device) {
-      if (device.connectionStatus == DeviceConnectionStatus.CONNECTED) {
+      if (device != null &&
+          device.connectionStatus == DeviceConnectionStatus.CONNECTED) {
         disconnectFromDevice(device);
       }
     });
@@ -43,7 +43,7 @@ class AppModel extends ChangeNotifier {
 
     try {
       Mds.startScan((name, address) {
-        Device device = Device(name, address);
+        Device device = Device(name ?? "", address ?? "");
         if (!_deviceList.contains(device)) {
           _deviceList.add(device);
           notifyListeners();
@@ -65,36 +65,42 @@ class AppModel extends ChangeNotifier {
 
   void connectToDevice(Device device) {
     device.onConnecting();
-    Mds.connect(device.address,
-            (serial) => _onDeviceMdsConnected(device.address, serial),
-            () => _onDeviceDisconnected(device.address),
-            () => _onDeviceConnectError(device.address)
-    );
+    final address = device.address;
+    if (address == null) return;
+    Mds.connect(
+        address,
+        (serial) => _onDeviceMdsConnected(address, serial),
+        () => _onDeviceDisconnected(address),
+        () => _onDeviceConnectError(address));
   }
 
   void disconnectFromDevice(Device device) {
-    Mds.disconnect(device.address);
-    _onDeviceDisconnected(device.address);
+    final address = device.address;
+    if (address == null) return;
+    Mds.disconnect(address);
+    _onDeviceDisconnected(address);
   }
 
   void _onDeviceMdsConnected(String address, String serial) {
-    Device foundDevice = _deviceList.firstWhere((element) => element.address == address);
+    Device? foundDevice =
+        _deviceList.firstWhere((element) => element?.address == address);
     if (foundDevice != null) {
       foundDevice.onMdsConnected(serial);
       notifyListeners();
       if (_onDeviceMdsConnectedCb != null) {
-        _onDeviceMdsConnectedCb.call(foundDevice);
+        _onDeviceMdsConnectedCb?.call(foundDevice);
       }
     }
   }
 
   void _onDeviceDisconnected(String address) {
-    Device foundDevice = _deviceList.firstWhere((element) => element.address == address);
+    Device? foundDevice =
+        _deviceList.firstWhere((element) => element?.address == address);
     if (foundDevice != null) {
       foundDevice.onDisconnected();
       notifyListeners();
       if (_onDeviceDisonnectedCb != null) {
-        _onDeviceDisonnectedCb.call(foundDevice);
+        _onDeviceDisonnectedCb?.call(foundDevice);
       }
     }
   }
