@@ -2,7 +2,7 @@
 //  Generated code. Do not modify.
 //  source: protos/mdsflutter.proto
 //
-// @dart = 2.3
+
 // ignore_for_file: annotate_overrides,camel_case_types,unnecessary_const,non_constant_identifier_names,library_prefixes,unused_import,unused_shown_name,return_of_invalid_type
 
 export 'mdsflutter.pb.dart';
